Add files via upload
o The Parameters section is used to prompt for inputs that can be used elsewhere in the template. The template is asking for two IP address (CIDR) ranges for defining the VPC.
o The Resources section is used to define the infrastructure to be deployed. The template is defining the VPC, and a Security Group.
o  The Outputs section is used to provide selective information about resources in the stack. The template is providing the Default Security Group for the VPC that is created.

The template is written in a format called YAML, which is commonly used for configuration files. This template will be used to launch a CloudFormation stack.