Add files via upload
This third yaml file includes the addition of an Amazon EC2 instance .
To create this file, the below document link was used as an reference:

https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-ec2-instance.html

The ImageId property specifies the Amazon Machine Image (AMI) that will be used to launch the instance. In this case, we are using the latest Amazon Linux 2 AMI, which is retrieved from the AWS Systems Manager Parameter Store using the AmazonLinuxAMIID parameter.

The InstanceType property specifies the instance type. In this case, we are using the t3.micro instance type, which is a small instance that is suitable for development and testing.

The SecurityGroupIds property specifies the security groups that the instance will be a member of. In this case, the instance will be a member of the AppSecurityGroup security group, which allows SSH access from the local machine and HTTP and HTTPS access from the internet.

The SubnetId property specifies the subnet that the instance will be launched into. In this case, the instance will be launched into the PublicSubnet subnet, which is a public subnet that has a route to the internet gateway.

The Tags property specifies the tags that will be applied to the instance. In this case, we are adding a tag with the key Name and the value App Server. This tag will make it easy to identify the instance in the AWS Management Console.